Fixes #8: recogineze quotes arround values in .env
Added a check to see if there are quotes (single or double) arround the value.
If there are the same kind of quotes around each side: remove them.

This commit also removes the usage of $1 and $2 variables for accessing the
backreferences because the added gsubs()'s overwrite those values. Named
captures are used instead.